+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/ornaments/attribution/MapAttributionScope.kt
@@ -46,7 +46,6 @@ import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.DialogProperties
 import com.mapbox.common.geofencing.GeofencingError
 import com.mapbox.maps.MapView
-import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.extension.compose.MapboxMapScopeMarker
 import com.mapbox.maps.extension.compose.R
 import com.mapbox.maps.extension.compose.ornaments.attribution.internal.AttributionComposePlugin
@@ -142,8 +141,6 @@ public class MapAttributionScope internal constructor(
    * @param telemetryDialog Defines TelemetryDialog when the Mapbox telemetry is clicked.
    * @param geofencingConsentDialog Defines GeofencingConsentDialog when the Mapbox Geofencing is clicked.
    */
-  @com.mapbox.annotation.MapboxExperimental
-  @MapboxExperimental
   @Composable
   public fun Attribution(
     modifier: Modifier = Modifier,